colocate BoxE methods
